Ideas
In this PR:
- docker stuff is about infrastructure
- logs are data
- requirements.txt contains the vendor and should be keep at the root
- changelog.txt should be at the root with readme.md, refactoring.md and so on.

----

*Api keys*
I think they should be in one file at the root of the repository, like: `user_config.example.json` with binance and bybit. This file could be given to the bot in the command line.

*Live settings file*
The magic behind getting the same file name in the live settings distub me, it could also be given through the CLI.

*Makefile*

Makefile is good way to give an unique entry point and durable contract for the user (like the docker-compose.yml file). It could be an idea to implement it like this:

```bash
.DEFAULT_GOAL: help

# List commands
help:
	@grep -E '^[a-zA-Z_-]+:.*?## .*$$' $(MAKEFILE_LIST) | awk 'BEGIN {FS = ":.*?## "}; {printf "\033[36m * %s\033[0m %s\n", $$1, $$2}'

run-live: ## Run bot with live settings. make run-live EXCHANGE=binance USERNAME=username
	python3 start_bot.py $(EXCHANGE) $(USERNAME)

run-live: ## Run bot with live settings. make run-live CONFIG=configname
	python3 backtest.py $(CONFIG)
```